# openapi.yml
openapi: 3.0.3
info:
  title: Reminder API
  description: Apple Reminder Web App REST API
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

tags:
  - name: lists
    description: 미리 알림 목록 관리

paths:
  /api/lists:
    get:
      tags: [lists]
      summary: 전체 목록 조회
      operationId: getAllLists
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ReminderListResponse'

    post:
      tags: [lists]
      summary: 목록 생성
      operationId: createList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderListRequest'
      responses:
        '201':
          description: 생성 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderListResponse'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /api/lists/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int64

    get:
      tags: [lists]
      summary: 목록 단건 조회
      operationId: getList
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderListResponse'
        '404':
          description: 목록을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    put:
      tags: [lists]
      summary: 목록 수정
      operationId: updateList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReminderListRequest'
      responses:
        '200':
          description: 수정 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReminderListResponse'
        '400':
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: 목록을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    delete:
      tags: [lists]
      summary: 목록 삭제
      operationId: deleteList
      responses:
        '204':
          description: 삭제 성공
        '404':
          description: 목록을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    ReminderListRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          description: 목록 이름
          example: 업무
        color:
          type: string
          description: 목록 색상 (hex)
          example: "#007AFF"
          nullable: true
        icon:
          type: string
          description: 목록 아이콘 (Lucide 아이콘명 또는 이모지)
          example: briefcase
          nullable: true
        groupId:
          type: integer
          format: int64
          description: 목록 그룹 ID
          nullable: true

    ReminderListResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 목록 ID
          example: 1
        name:
          type: string
          description: 목록 이름
          example: 업무
        color:
          type: string
          description: 목록 색상 (hex)
          example: "#007AFF"
          nullable: true
        icon:
          type: string
          description: 목록 아이콘
          example: briefcase
          nullable: true
        groupId:
          type: integer
          format: int64
          description: 목록 그룹 ID
          nullable: true
        sortOrder:
          type: integer
          description: 정렬 순서
          example: 0
        createdAt:
          type: string
          format: date-time
          description: 생성 일시
          example: "2026-03-31T00:00:00"

    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
          description: HTTP 상태 코드
          example: 404
        message:
          type: string
          description: 오류 메시지
          example: "ReminderList not found: 999"